evaluation improved, cleaning up in tests

+++ src/com/strategy/api/logic/evaluation/Evaluation.java
@@ -11,6 +11,6 @@ import com.strategy.api.logic.situation.Situation;
  */
 public interface Evaluation {
 
-	Map<Integer, Integer> getEvaluatedFields();
+	Map<Integer, Double> getEvaluatedFields();
 
 }
